pyproject.toml
[tool.poetry]
authors = ["Philip Moraski <ptmoraski@gmail.com>"]
description = "A Household Chore Manager."
package-mode = false
license = "AGPL"
name = "GopherDen"
version = "0.0.1a"


[tool.poetry.scripts]
start = "backend.app:main"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
python-slugify = "^8.0.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
sqlalchemy = "^2.0.31"
bcrypt = "^4.0.1"
pyhumps = "^3.5.3"
text-unidecode = "^1.3"
requests = "^2.32.3"

[tool.pytest.ini_options]
addopts = "-ra -q"
minversion = "6.0"
python_classes = '*Tests'
python_files = 'test_*'
python_functions = 'test_*'
testpaths = ["tests"]